change schema data type for batch analyze

+++ app/models/batch_analyze_model.py
@@ -62,7 +62,7 @@ class CandidateAnalysisResponse(BaseModel):
     email: Optional[str] = None
     phone: Optional[str] = None
     currentTitle: Optional[str] = None
-    experienceYears: Optional[int] = None
+    experienceYears: Optional[float] = None
     skills: Optional[List[dict]] = []
     availability: Optional[str] = None
     matchScore: Optional[float] = 0.0
